# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "vms Api Documentation",
    "description": "vms Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "vms Logo",
      "altText": "vms Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "vms API Support",
      "url": "https://vms.com/info/support",
      "email": "support@vms.com"
    },
    "termsOfService": "https://vms.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/reparation": {
      "get": {
        "summary": "List Reparation",
        "description": "Reparation",
        "operationId": "ListReparation",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Reparation"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListReparation"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/reparation/view/{id}": {
      "get": {
        "summary": "View Reparation",
        "description": "Vue Reparation",
        "operationId": "ViewReparation",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Reparation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewReparation"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/reparation/add": {
      "post": {
        "summary": "Add Reparation",
        "description": "Ajouter un nouveau",
        "operationId": "AddReparation",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Reparation"
        ],
        
        "requestBody": {
          "description": "Add Reparation",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReparation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewReparation"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/reparation/edit/{id}": {
      "put": {
        "summary": "Edit Reparation",
        "description": "modifier",
        "operationId": "EditReparation",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Reparation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Reparation",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditReparation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditReparation"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/reparation/delete/{id}": {
      "delete": {
        "summary": "Delete Reparation",
        "description": "Reparation",
        "operationId": "DeleteReparation",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Reparation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ListReparation": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          },
          "date_created": {
            "description": "Date Created",
            "type": "string"
          },
          "date_updated": {
            "description": "Date Updated",
            "type": "string"
          },
          "user_id": {
            "description": "User Id",
            "type": "string"
          }
        }
      },
      "ViewReparation": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          },
          "date_created": {
            "description": "Date Created",
            "type": "string"
          },
          "date_updated": {
            "description": "Date Updated",
            "type": "string"
          },
          "user_id": {
            "description": "User Id",
            "type": "string"
          }
        }
      },
      "AddReparation": {
        "type": "object",
        "properties": {
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          },
          "user_id": {
            "description": "User Id",
            "type": "string"
          }
        }
      },
      "EditReparation": {
        "type": "object",
        "properties": {
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          },
          "user_id": {
            "description": "User Id",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}